modified:   backend/app.sqlite 	modified:   package-lock.json 	modified:   package.json 	modified:   src/app.jsx 	deleted:    src/assets/aula.jpg 	deleted:    src/assets/logo.png 	deleted:    src/assets/react.svg 	modified:   src/components/bars/navbar.jsx 	modified:   src/components/bars/sidebar-container.jsx 	modified:   src/components/pages/courses.jsx 	modified:   src/components/pages/courses_cards.jsx 	new file:   src/components/pages/courses_list.jsx 	modified:   src/components/pages/enroll-courses.jsx 	modified:   src/components/pages/my-task.jsx 	modified:   src/components/student-progress/quiz.jsx 	modified:   src/style/App.css 	modified:   src/style/dashboard.scss 	modified:   src/style/navbar.scss 	modified:   src/style/sidebar.scss 	modified:   src/style/tasks.scss

+++ src/components/pages/courses_cards.jsx
@@ -6,7 +6,6 @@ import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
 import { Button, CardActionArea, CardActions } from '@mui/material';
-import aula from '../../assets/aula.jpg'
 import { useNavigate } from "react-router-dom";
 
 
@@ -14,6 +13,14 @@ export default function CourseCards() {
   
   const [myCourses, setMyCourses] = useState([]);
   const nav = useNavigate();
+  const UnsplashImages = [
+    'https://images.unsplash.com/photo-1547394765-185e1e68f34e?auto=format&fit=crop&q=80&w=2070&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+    'https://images.unsplash.com/photo-1552664730-d307ca884978?auto=format&fit=crop&q=80&w=2070&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D/fotos/mujer-colocando-notas-adhesivas-en-la-pared-Oalh2MojUuk',
+    'https://images.unsplash.com/photo-1551836022-d5d88e9218df?auto=format&fit=crop&q=80&w=2070&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+    'https://images.unsplash.com/photo-1515378960530-7c0da6231fb1?auto=format&fit=crop&q=80&w=2070&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+    'https://images.unsplash.com/photo-1491975474562-1f4e30bc9468?auto=format&fit=crop&q=80&w=1974&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D'
+  ];
+  
 
   const setId = (courseId) => {
     localStorage.setItem('courseId', courseId)
@@ -23,7 +30,8 @@ export default function CourseCards() {
  useEffect(() => {
     getStudentCourses();
   }, []);
-   
+
+    
   async function getStudentCourses() {        
         try{
         const { data: { user } } = await supabase.auth.getUser()     
@@ -36,30 +44,34 @@ export default function CourseCards() {
       }};  
        
   async function dropCourse(courseId) {        
-    try{
-    const { data: { user } } = await supabase.auth.getUser()     
-    const {data: error} = await supabase.from('dasboard')
-        .delete().match({student_id: user.id, course_id: courseId}); 
+    const confirmation = window.confirm("¿Estás seguro de que deseas eliminar este curso?"); // Muestra una alerta y espera la respuesta del usuario
+
+    if (confirmation) {
+      try {
+        const { data: { user } } = await supabase.auth.getUser();
+        const { data: error } = await supabase.from('dasboard')
+          .delete().match({ student_id: user.id, course_id: courseId });
         getStudentCourses();
         if (error) {
           console.log(error);
         }
       } catch (error) {
         console.log('error', error);
-      }}; 
+      }
+    }}; 
 
   return (
     <div className="card-container">
 
-    {myCourses.map((course) => (
+    {myCourses.map((course, index) => (
 
     <Card sx={{ Width: 340 , minWidth: 320, m: 3 }} key={course.id}>
       <CardActionArea onClick={() => setId(course.id)}>
         <CardMedia
           component="img"
           height="140"
-          image={aula}
-          alt="aula"
+          image={UnsplashImages[index]}
+          alt={course.name}
         />
         <CardContent >
           <Typography gutterBottom variant="h5" component="div" key={course.name}>
@@ -69,7 +81,7 @@ export default function CourseCards() {
           {course.description}
           </Typography>
           <Typography variant="body2" color="text.secondary" key={course.hours}>
-          {course.hours}
+          {course.hours} hours
           </Typography>
         </CardContent>
       </CardActionArea>
@@ -83,22 +95,3 @@ export default function CourseCards() {
    </div>
 );
 }
-
-
-
-    {/* 
-     
-      {myCourses.map((course) => (
-        <table className='courses-table' key={course.id}>
-            
-                  <td key={course.name}>{course.name}</td>
-                  <td key={course.hours}>{course.hours}</td>
-                  <td key={course.description}>{course.description}</td>
-                  <td className="operation">
-                    <button className="btn-drop" onClick={() => dropCourse(course.id)}>Drop</button>
-                  </td>
-   
-       
-          ))} 
-    </div>
-    </div> */}
